+++ src/main/java/com/contentful/java/cma/CMAClient.java
@@ -67,6 +67,8 @@ public class CMAClient {
   private final ModuleEntries moduleEntries;
   private final ModuleSpaces moduleSpaces;
   private final ModuleWebhooks moduleWebhooks;
+  private final ModuleUsers moduleUsers;
+  private final ModuleSpaceMemberships moduleSpaceMemberships;
   private final ModuleUploads moduleUploads;
 
   // Executors
@@ -110,6 +112,8 @@ public class CMAClient {
     this.moduleEntries = new ModuleEntries(retrofit, callbackExecutor);
     this.moduleSpaces = new ModuleSpaces(retrofit, callbackExecutor);
     this.moduleWebhooks = new ModuleWebhooks(retrofit, callbackExecutor);
+    this.moduleUsers = new ModuleUsers(retrofit, callbackExecutor);
+    this.moduleSpaceMemberships = new ModuleSpaceMemberships(retrofit, callbackExecutor);
 
     this.moduleUploads = new ModuleUploads(uploadRetrofit, callbackExecutor);
   }
@@ -183,6 +187,20 @@ public class CMAClient {
    */
   public ModuleWebhooks webhooks() {
     return moduleWebhooks;
+  }
+
+  /**
+   * @return the Users module.
+   */
+  public ModuleUsers users() {
+    return moduleUsers;
+  }
+
+  /**
+   * @return the SpaceMembership module.
+   */
+  public ModuleSpaceMemberships spaceMemberships() {
+    return moduleSpaceMemberships;
   }
 
   /**